Improve class responsibilities and docs
